[FIXED] Parsing File Bug
This problem was caused because of "isMnemonic()" function! It always returned TRUE for the first token in each line, so the program treated the first 3 lines as assembly instructions!

See the following assembly code:

The first 3 lines should not be included in a assembly code region:

hinstDLL	= dword	ptr  4
fdwReason	= dword	ptr  8
lpReserved	= dword	ptr  0Ch

	push	1
	pop	eax
	retn	0Ch