New Unification Procedure:
Now, the accepted offset to be considered for clone overlap is: (1- minCoOccurThr) * windowSize.
i.e.:
minCoOccurTresh = 0.8
windowSize = 10

Accepted offset is : 2
It means we can have two code fragments with size 10 and 12 as clones!
If we increase the threshold, the accepted offset will be decreased and vice versa. Logical enough to be defended!